# [WIP] Avoid unnecessary split for degenerate case where all labels are identical, by checking label range