Partially addresses #3188 by not adding a new split for the special case where the data in the node have all identical labels. This degenerate case may occur if the data was generated with `make_blobs`.

Closes #3128