@hcho3 2020-12-02
Designating as WIP, since this change appears to result into suboptimal choice of split as in the case of https://github.com/rapidsai/cuml/issues/3188#issue-750663748