@hcho3 2020-12-02
The suboptimal split was caused by incorrect choice of CUDA stream in the `initLabelRange` kernel. I pushed the fix.