@hcho3 2020-12-02
This PR also fixes #3128.

```
python -m cuml.benchmark.run_benchmarks --cuml-param-sweep use_experimental_backend=[true,false]  \
  max_depth=[10,30] --n-reps 1 --skip-cpu --num-rows 1000000 --num-features 256  \
  RandomForestRegressor
```

Before:
```
                     algo  input     cu_time  cpu_time  cuml_acc  cpu_acc  speedup  n_samples  n_features  use_experimental_backend  max_depth
0   RandomForestRegressor  numpy   11.271957       0.0  0.999997      0.0      0.0    1000000         256                      True         10
1   RandomForestRegressor  numpy  115.319010       0.0  0.991797      0.0      0.0    1000000         256                      True         30
2   RandomForestRegressor  numpy    2.103713       0.0  0.999999      0.0      0.0    1000000         256                     False         10
3   RandomForestRegressor  numpy    3.486071       0.0  0.999999      0.0      0.0    1000000         256                     False         30
```

After:
```
                    algo  input   cu_time  cpu_time  cuml_acc  cpu_acc  speedup  n_samples  n_features  use_experimental_backend  max_depth
0  RandomForestRegressor  numpy  7.107988       0.0  0.999999      0.0      0.0    1000000         256                      True         10
1  RandomForestRegressor  numpy  7.835434       0.0  0.999999      0.0      0.0    1000000         256                      True         30
2  RandomForestRegressor  numpy  2.082379       0.0  0.999999      0.0      0.0    1000000         256                     False         10
3  RandomForestRegressor  numpy  3.591269       0.0  0.999999      0.0      0.0    1000000         256                     False         30
```